Update folder structure and move teleport cursor from core to loco package
