+++ Assets~/StandardAssets/Cursors.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 092e8a30d43c85a4bbbe580f6ae971df
+guid: f1f384cb3af0ecb4a9272d25245b14a1
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
