+++ Assets~/StandardAssets/Cursors/TeleportCursor.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 90690583c223e3d43b1c2cb641a3efdc
-PrefabImporter:
+guid: 0a8f46d0b2bb8ea47b98329384d3ae6c
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
