+++ Assets~/StandardAssets/Cursors/TeleportCursor/TeleportCursor.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dbcafce56f49455588eb636dbf45d3e6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
