+++ Assets~/StandardAssets/Cursors/TeleportCursor/TeleportCursorController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 382528586dea4be993a8c1d11a635a32
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
