+++ Assets~/StandardAssets/Cursors/TeleportCursor/TeleportCursorDisabled.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 985b9d3698994b088dc4d9fb090216cb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
