+++ Assets~/StandardAssets/Cursors/TeleportCursor/TeleportInvalid.anim.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 9dca11c400d1e9342ac7e9ca4bd6e83b
+guid: e25db01aca62498488ef94185bf22550
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
