+++ Assets~/StandardAssets/Cursors/TeleportCursor/TeleportValid.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a0e1d9113fe48459abfb30a984b4a8f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
