+++ Assets~/StandardAssets/Interactors.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cce747f0ca3508d48b8cacc05b2bbc7e
+guid: 2a78b204a12b5e24b93a4b1515a7600e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
