+++ Assets~/StandardAssets/TeleportAnchor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 78729aebf986a1640b344a669bd06ea5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
