+++ Assets~/StandardAssets/TeleportAnchor/TeleportAnchor_OverrideOrientation.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 72a36ac6f2522d149809e34c4663e77a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
