--- Assets~/StandardAssets/TeleportAnchor_WithOrientationOverride.prefab
@@ -1,63 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &324964587421388291
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 3396248214911929764, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_Name
-      value: TeleportAnchor_WithOrientation
-      objectReference: {fileID: 0}
-    - target: {fileID: 8598909145488103567, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: overrideOrientation
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9085342681829300181, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 10c301cea4d381145a6ccc7f595a6dd3, type: 3}
