+++ Runtime/Cursors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 67fedd34cc68cf741900cc0d860d40e3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
